Update how-to-get-attachments-by-using-ews-in-exchange.md
The Attachment.Load() is an async task that needs an await or Wait();
The Attachment.Load(string fileName) calls the Load() method internally. As far as I can tell, it does not await this task. That means that only empty files are saved and the content is not copied into the new file